Fix use-after-delete for Celeritas Geant4 exception handler
