+++ src/celeritas/CMakeLists.txt
@@ -120,8 +120,8 @@ if(CELERITAS_USE_Geant4)
     ext/GeantImporter.cc
     ext/GeantSetup.cc
     ext/GeantVolumeMapper.cc
+    ext/ScopedGeantExceptionHandler.cc
     ext/detail/GeantBremsstrahlungProcess.cc
-    ext/detail/GeantExceptionHandler.cc
     ext/detail/GeantLoggerAdapter.cc
     ext/detail/GeantMicroXsCalculator.cc
     ext/detail/GeantModelImporter.cc
